Obfuscated C code misinterpretation data

Data from Gopstein et. al.'s experiment on the misinterpretation of C code. Subjects were asked to hand evaluate pairs of functionally equivalent code. Half of the questions were intentionally obfuscated to elicit confusion.

data(obfuscation)

Format

A data frame with 57 rows and 4 variables:

subject

the ID of the study participant

atom

the type of obfuscation being evaluated

control

whether the subject answered the un-obfuscated question correctly

treatment

whether the subject answered the obfuscated question correctly

Source

http://atomsofconfusion.com

Examples

data(obfuscation)

oc <- paired.to.contingency(group = obfuscation[,c("subject", "atom")],
                            t1    = obfuscation$control,
                            t2    = obfuscation$treatment)

clust.bin.pair(oc$ak, oc$bk, oc$ck, oc$dk, method="durkalski")

#> 
#> 	Durkalski's Chi-square test
#> 
#> data:  oc$ak, oc$bk, oc$ck, oc$dk
#> chi-square = 10.125, df = 1, p-value = 0.001463
#> 
#>