Convert between paired results and the canonical contingency tables

Group results by common clustering then tally the concordant and discordant pairs.

paired.to.contingency(group, t1, t2)

Arguments

group

List of grouping values

t1

pre-treatment measures

t2

post-treatment measures

Value

Contingency tables represented in the rows of a matrix

Examples

paired.to.contingency(list(obfuscation$subject, obfuscation$atom),
                      obfuscation$control, obfuscation$treatment)

#>    Group.1      Group.2 ak bk ck dk
#> 1        1    CONDITION  0  0  1  1
#> 2        2    CONDITION  1  0  1  0
#> 3        3    CONDITION  0  0  0  2
#> 4        4    CONDITION  0  0  1  1
#> 5        6    CONDITION  0  0  1  1
#> 6        7    CONDITION  2  0  0  0
#> 7        8    CONDITION  2  0  0  0
#> 8       10    CONDITION  2  0  0  0
#> 9       11    CONDITION  2  0  0  0
#> 10      12    CONDITION  1  0  0  1
#> 11      13    CONDITION  2  0  0  0
#> 12      14    CONDITION  2  0  0  0
#> 13      15    CONDITION  1  0  1  0
#> 14      16    CONDITION  1  0  1  0
#> 15      17    CONDITION  2  0  0  0
#> 16      18    CONDITION  2  0  0  0
#> 17      19    CONDITION  1  0  1  0
#> 18       1  PARENTHESIS  1  1  0  0
#> 19       2  PARENTHESIS  1  0  0  1
#> 20       3  PARENTHESIS  0  0  0  2
#> 21       4  PARENTHESIS  1  1  0  0
#> 22       6  PARENTHESIS  0  0  1  1
#> 23       7  PARENTHESIS  2  0  0  0
#> 24       8  PARENTHESIS  1  0  1  0
#> 25      10  PARENTHESIS  2  0  0  0
#> 26      11  PARENTHESIS  2  0  0  0
#> 27      12  PARENTHESIS  2  0  0  0
#> 28      13  PARENTHESIS  2  0  0  0
#> 29      14  PARENTHESIS  2  0  0  0
#> 30      15  PARENTHESIS  2  0  0  0
#> 31      16  PARENTHESIS  2  0  0  0
#> 32      17  PARENTHESIS  2  0  0  0
#> 33      18  PARENTHESIS  1  0  1  0
#> 34      19  PARENTHESIS  1  0  1  0
#> 35       1 POST_INC_DEC  0  1  0  1
#> 36       2 POST_INC_DEC  1  0  1  0
#> 37       3 POST_INC_DEC  0  0  2  0
#> 38       4 POST_INC_DEC  0  0  0  2
#> 39       6 POST_INC_DEC  0  0  0  2
#> 40       7 POST_INC_DEC  2  0  0  0
#> 41       8 POST_INC_DEC  2  0  0  0
#> 42      10 POST_INC_DEC  2  0  0  0
#> 43      11 POST_INC_DEC  2  0  0  0
#> 44      12 POST_INC_DEC  2  0  0  0
#> 45      13 POST_INC_DEC  2  0  0  0
#> 46      14 POST_INC_DEC  2  0  0  0
#> 47      15 POST_INC_DEC  2  0  0  0
#> 48      16 POST_INC_DEC  2  0  0  0
#> 49      17 POST_INC_DEC  2  0  0  0
#> 50      18 POST_INC_DEC  1  0  1  0
#> 51      19 POST_INC_DEC  0  1  1  0
#> 52       1  PRE_INC_DEC  0  0  1  0
#> 53       2  PRE_INC_DEC  1  0  0  0
#> 54       4  PRE_INC_DEC  0  0  0  1
#> 55       7  PRE_INC_DEC  1  0  0  0
#> 56       8  PRE_INC_DEC  0  0  1  0
#> 57      10  PRE_INC_DEC  1  0  0  0
#> 58      11  PRE_INC_DEC  1  0  0  0
#> 59      13  PRE_INC_DEC  1  0  0  0
#> 60      14  PRE_INC_DEC  1  0  0  0
#> 61      16  PRE_INC_DEC  1  0  0  0
#> 62      18  PRE_INC_DEC  0  0  1  0
#> 63      19  PRE_INC_DEC  0  0  0  1
#>