Psychiatrist and patient disagreement data

Psychiatrists and their patients were surveyed in pairs regarding patient concerns and treatment. Each psychiatrist was asked whether each question item was relevant to their patient and each of their patients were asked the same. The data can be evaluated to answer the question of whether there was patient/doctor agreement in each item. The sample was 29 psychiatrists, each with 1-8 patients, for a total of N = 135 matched pairs.

data(psychiatry)

Format

A data frame with 29 rows and 7 variables:

psychiatrist

the ID of the psychiatrist

Nh

the number of the psychiatrist's patients participating in the experiment

ah

both participants answered 1

bh

patient answered 1, psychiatrist answered 0

ch

patient answered 0, psychiatrist answered 1

dh

both participants answered 0

Wh

Normalized difference: (bh - ch) / Nh

Source

Donner, A., & Petryshen, P. (1989). The statistical analysis of matched data in psychiatric research. Psychiatry research, 28(1), 41-46.

References

Eliasziw, M., & Donner, A. (1991). Application of the McNemar test to non‐independent matched pair data. Statistics in medicine, 10(12), 1981-1991.

Examples

data(psychiatry)

psychiatry$Wh == round((psychiatry$bh - psychiatry$ch) / psychiatry$Nh, 2)

#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE
#> [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [25]  TRUE  TRUE  TRUE  TRUE  TRUE
#> 

clust.bin.pair(psychiatry$ah, psychiatry$bh, psychiatry$ch, psychiatry$dh, method="eliasziw")

#> 
#> 	Eliasziw's Chi-square test
#> 
#> data:  psychiatry$ah, psychiatry$bh, psychiatry$ch, psychiatry$dh
#> chi-square = 10.226, df = 1, p-value = 0.001385
#> 
#>