PET and SPECT data for diagnosing hyperparathyroidism

Following surgery which confirmed the absence of hyperparathyroidism two diagnostic tests, PET and SPECT, were performed. Their measures of true negatives and false positives are reported. Data reported in Obuchowki 1998.

data(thyroids)

Format

A data frame with 21 rows and 6 variables:

patient

ID of the patient

n.glands

number of glands tested from the patient

n.pet

number of true negatives from the PET test

x.pet

individual results per gland from the PET test

n.spect

number of true negatives from the SPECT test

x.spect

individual results per gland from the SPECT test

Source

Obuchowski, N. A. (1998). On the comparison of correlated proportions for clustered data. Statistics in medicine, 17(13), 1495-1507.

Examples

data(thyroids)

thyroids$n.glands == sapply(thyroids$x.pet, length)

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [16] TRUE TRUE TRUE TRUE TRUE TRUE
#> 

thyroids$n.glands == sapply(thyroids$x.spect, length)

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [16] TRUE TRUE TRUE TRUE TRUE TRUE
#> 

thyroids$n.pet   == sapply(thyroids$x.pet,   function(x) length(which(x == 1)))

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [16] TRUE TRUE TRUE TRUE TRUE TRUE
#> 

thyroids$n.spect == sapply(thyroids$x.spect, function(x) length(which(x == 1)))

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [16] TRUE TRUE TRUE TRUE TRUE TRUE
#> 

tc <- nested.to.contingency(thyroids$x.pet, thyroids$x.spect)

clust.bin.pair(tc[,'ak'], tc[,'bk'], tc[,'ck'], tc[,'dk'], method="obuchowski")

#> 
#> 	Obuchowski's Chi-square test
#> 
#> data:  tc[, "ak"], tc[, "bk"], tc[, "ck"], tc[, "dk"]
#> chi-square = 2.8571, df = 1, p-value = 0.09097
#> 
#> 

do.call(clust.bin.pair, data.frame(tc))

#> 
#> 	Yang's Chi-square test
#> 
#> data:  c("c(0, 2, 3, 1, 2, 4, 3, 2, 1, 1, 2, 2, 3, 2, 0, 2, 2, 2, 2, 1, ", "2)"), c("c(0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ", "0)"), c("c(2, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 1, 0, 0, ", "0)"), c("c(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, ", "0)")
#> chi-square = 3.1317, df = 1, p-value = 0.07679
#> 
#>